+++ havefnubb/modules/hfnuadmin/controllers/ranks.classic.php
@@ -13,22 +13,28 @@ class ranksCtrl extends jController {
     *
     */
     public $pluginParams = array(
-        'create'    => array( 'jacl2.right'=>'hfnu.admin.rank.create'),
-        'edit'      => array( 'jacl2.right'=>'hfnu.admin.rank.edit'),
+        'index'    => array( 'jacl2.rights.and'=>array('hfnu.admin.rank.create',
+														'hfnu.admin.rank.edit')
+							),		
         'delete'    => array( 'jacl2.right'=>'hfnu.admin.rank.delete'),
     );
     
-
-    function create () {
-        $form = jForms::create('hfnuadmin~ranks');
-        $rep = $this->getResponse('html');
-        $tpl = new jTpl();
+	function index() {
+		$tpl = new jTpl();
+		
+		$form = jForms::create('hfnuadmin~ranks');
+		
+        $dao = jDao::get('havefnubb~ranks');
+        $ranks = $dao->findAll();
+		
         $tpl->assign('form', $form);
-        $tpl->assign('action','hfnuadmin~ranks:savecreate');
-        $tpl->assign('rank_heading',jLocale::get('hfnuadmin~rank.create.a.rank'));
-        $rep->body->assign('MAIN',$tpl->fetch('ranks_edit'));
-        return $rep;        
-    }
+        $tpl->assign('ranks',$ranks);
+				
+        $rep = $this->getResponse('html');        
+        $rep->body->assign('MAIN', $tpl->fetch('hfnuadmin~ranks_index'));
+        return $rep;
+	}
+
 
     function savecreate () {
         $form = jForms::get('hfnuadmin~ranks');
@@ -52,39 +58,12 @@ class ranksCtrl extends jController {
             $dao->insert($record);
             
             jMessage::add(jLocale::get('hfnuadmin~rank.rank.added'),'ok');
-            $rep = $this->getResponse('redirect');
-            $rep->action='hfnuadmin~default:ranks';
-            return $rep;
         }
-        else {
-            $rep = $this->getResponse('redirect');
-            $rep->action='hfnuadmin~default:ranks';
-            return $rep;
-        }        
-    }
-
-    function edit () {
-        $id_rank = (int) $this->param('id_rank');
-        
-        if ($id_rank == 0) {
-            jMessage::add(jLocale::get('hfnuadmin~rank.unknown.rank'),'error');
-            $rep = $this->getResponse('redirect');
-            $rep->action='hfnuadmin~default:rank';
-            return $rep;                 
-        }      
-
-		$form = jForms::create('hfnuadmin~ranks',$id_rank);
-		$form->initFromDao("havefnubb~ranks");
-				
-		$form->setData('id_rank',$id_rank);
-            
-        $rep = $this->getResponse('html');
-        $tpl = new jTpl();
-        $tpl->assign('form', $form);
-        $tpl->assign('action','hfnuadmin~ranks:saveedit');
-        $tpl->assign('rank_heading',jLocale::get('hfnuadmin~rank.edit.this.rank'));
-        $rep->body->assign('MAIN',$tpl->fetch('ranks_edit'));
-        return $rep;        
+		
+		$rep = $this->getResponse('redirect');
+		$rep->action='hfnuadmin~ranks:index';		
+        return $rep;
+	    
     }
 
     function saveedit () {
@@ -93,7 +72,7 @@ class ranksCtrl extends jController {
         if ($id_rank == 0) {
             jMessage::add(jLocale::get('hfnuadmin~rank.unknown.rank'),'error');
             $rep = $this->getResponse('redirect');
-            $rep->action='hfnuadmin~default:rank';
+            $rep->action='hfnuadmin~ranks:index';
             return $rep;                 
         } 
         
@@ -102,7 +81,7 @@ class ranksCtrl extends jController {
         if (!$form->check()) {
             jMessage::add(jLocale::get('hfnuadmin~rank.invalid.datas'),'error');
             $rep = $this->getResponse('redirect');
-            $rep->action='hfnuadmin~default:ranks';
+            $rep->action='hfnuadmin~ranks:index';
             return $rep;
         }
         
@@ -115,20 +94,24 @@ class ranksCtrl extends jController {
         $dao->update($record);
         jMessage::add(jLocale::get('hfnuadmin~rank.rank.modified'),'ok');
         $rep = $this->getResponse('redirect');
-        $rep->action='hfnuadmin~default:ranks';
+        $rep->action='hfnuadmin~ranks:index';
         return $rep;
     }
 
     function delete() {
 		$id_rank = (integer) $this->param('id_rank');
-	
-		$dao = jDao::get('havefnubb~ranks');        
-        $dao->delete($id_rank);
-        
-        jMessage::add(jLocale::get('hfnuadmin~rank.rank.deleted'),'ok');
+		if ($id_rank == 0) {
+			jMessage::add(jLocale::get('hfnuadmin~rank.invalid.datas'),'error');			
+		} else {
+			$dao = jDao::get('havefnubb~ranks');
+			//@TODO : check if there is existing Member using this rank
+			// if so, tell the admin he cant remove it !		
+			$dao->delete($id_rank);        
+			jMessage::add(jLocale::get('hfnuadmin~rank.rank.deleted'),'ok');			
+		}
         
         $rep = $this->getResponse('redirect');
-        $rep->action='hfnuadmin~default:ranks';
+        $rep->action='hfnuadmin~ranks:index';
         return $rep;        
     }
